[#26] reducer: untie from the broker version
The reducer doesn't have to be tied to a specific broker version. When
we deploy a broker we're always relying on the operator to chose the
latest available version for us. Therefore the reducer should be
compatible with this.

However that's not to say that we can't "gate" UX features behind a flag
or something. But that would've to be independent.