[#44] broker creation: enabling rbac
This commit enables a RBAC control for the broker. When the user decides
to lock down the acccess to the broker using tokens, it has for side
effect to make the broker require a login to communicate with acceptors.
In this commit we make sure that the user is able to create and
configure users so that the acceptors are always accessible.