[#22] overhauling cert generation
Serving the console over tls means that the jolokia api server needs to
trust the certificate used by the console. The chosen technique to
perform this trust check to rely on the issuer CA.

This means that now the console needs to use a certificate issued by a
known issuer to the jolokia api server and that's where this change
comes into play.

The cert generation routine is now actively requesting the user to
provide a cluster issuer in order to fulfil its duties.