[NO-ISSUE] upgrade to openshift console 20
This upgrades the package.json to match openshift console 20.
It upgrades the necessary dependencies linked to that as well.
The most notable change is an upgrade of the typescript version.
The code also cleans up old typescript related issues such as types
never really declared.
Tests warning are fixed as well.