[#161] restricted mode: a new switch
Adding a physical switch to enable restricted mode. When navigating in
between restricted and non restricted users loose all their prior
configuration. This is because the two modes are fundamentally
different.

We do use the same mechanism to route the reducer as in between 7.12 and
7.13. The dependency is 7.12 <- 7.13 <- Restricted. This means that we
can still reuse pieces of config from 7.13 and 7.12 for the restricted
one. However it is yet unclear if that would be a pro or a con.

For the time being it allows easily switching between modes.